Implements Controller.
