+++ src/feature/Task/Patch/StyledComponents.ts
@@ -1,5 +1,5 @@
 import { Form } from "react-router-dom";
-import { styled } from "styled-components";
+import styled from "styled-components";
 
 export const FormWrapper = styled(Form)`
   display: flex;
