+++ src/feature/Task/utils/Utils.ts
@@ -35,7 +35,7 @@ export function GetOptions(baseUrl: string): Option[] {
 }
 
 export async function postData<T>(baseUrl: string, headers: HeadersInit, bodyData: T) {
-  return fetch(baseUrl, {
+  return await fetch(baseUrl, {
     method: 'POST',
     headers: headers,
     body: JSON.stringify(bodyData)
